# perf: use arrays for low branching runes