--- a/Taskfile.yaml
+++ b/Taskfile.yaml
@@ -1,19 +1,19 @@
 # yaml-language-server: $schema=https://taskfile.dev/schema.json
-version: "3"
+version: '3'
 
 tasks:
   test:
     cmds:
       - go test ./... {{.CLI_ARGS}}
-  
+
   bench:
     cmds:
-      - go test -bench=. ./... 
-    
+      - go test -bench=. -benchmem ./...
+
   generate:
     cmds:
       - go run ./scripts/main.go
-  
+
   upgrade:
     cmds:
-      - go get -u ./... 
+      - go get -u ./...