--- a/bench_test.go
+++ b/bench_test.go
@@ -1,7 +1,6 @@
 package useragent_test
 
 import (
-	"strings"
 	"testing"
 
 	ua "github.com/medama-io/go-useragent"
@@ -12,14 +11,13 @@ var result ua.UserAgent
 func BenchmarkParserAll(b *testing.B) {
 	parser := ua.NewParser()
 
-	for _, k := range testCases {
-		name := strings.ReplaceAll(k, " ", "")
-		b.Run(name, func(b *testing.B) {
+	b.Run("All", func(b *testing.B) {
+		for _, k := range testCases {
 			for i := 0; i < b.N; i++ {
 				result = parser.Parse(k)
 			}
-		})
-	}
+		}
+	})
 }
 
 func BenchmarkParserSingle(b *testing.B) {
@@ -27,5 +25,19 @@ func BenchmarkParserSingle(b *testing.B) {
 
 	for i := 0; i < b.N; i++ {
 		result = parser.Parse(testCases[0])
+	}
+}
+
+func BenchmarkParserPutAll(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		ua.NewParser()
+	}
+}
+
+func BenchmarkParserPutSingle(b *testing.B) {
+	trie := ua.NewRuneTrie()
+
+	for i := 0; i < b.N; i++ {
+		trie.Put(testCases[0])
 	}
 }