This reduces the number of maps in the final trie structure by relying on arrays with a linear search for runes with less than 64 children. 

Old:
```
BenchmarkParserAll/All-16          61072             19564 ns/op             576 B/op         55 allocs/op
BenchmarkParserSingle-16         1996647               518.8 ns/op            31 B/op          2 allocs/op
BenchmarkParserPutAll-16              64          16484833 ns/op        15168293 B/op     248607 allocs/op
BenchmarkParserPutSingle-16       405416              2707 ns/op             448 B/op         12 allocs/op
```

New:
```
BenchmarkParserAll/All-16          69675             16349 ns/op             576 B/op         55 allocs/op
BenchmarkParserSingle-16         2621797               440.9 ns/op            26 B/op          2 allocs/op
BenchmarkParserPutAll-16              93          12369175 ns/op         7014044 B/op     176055 allocs/op
BenchmarkParserPutSingle-16       438016              2609 ns/op             448 B/op         12 allocs/op
```