@ayuhito 2024-09-04
This also brought down the runtime memory usage from around ~18-19MB to ~7MB.